Added MiniMap to WorldView and CellView.
For WorldView, a polygon representing the bounds of the world is shown.
This could be improved by indicating non-empty cells.

For CellView, the map/Lot images are shown.

Unfortunately, scrolling is slowed way down when the MiniMap is visible.
I think this is only because the mini-map obscures part of the scene, not
because the mini-map is itself slow to render.