MapImageManager now correctly handles out-of-bounds Lots.
When a Lot was outside a map's bounds (except off the top of the map), it
wasn't possible to determine tile coordinates inside the map image correctly
due to the map being offset within the image by the Lot.  Previously, it was
taken for granted that the width of the image was equal to the width of the
map.

The MapImageManager now maintains a binary blob (one .dat file per thumbnail)
that records enough information to properly determine tile coordinates in a
map image.