No more converting map orientations just to display them.
The MapManager had lots of fugly code to convert maps from Isometric to
LevelIsometric and vice-versa just so maps with different orientations
could be displayed together (i.e., Lots in a parent map).

Now MapComposite and CompositeLayerGroup do all the dirty work of calculating
the adjusted position of tiles when a maps's orientation differs from the
MapRenderer's orientation.