Allow objects to be reordered via drag-and-drop.
WorldCellObjects could already be moved to a new group or level via dnd,
but that did not affect their order in the cell's list of objects.

CellScene now sets the z-order of ObjectItems to match that in the ObjectsDock,
sorting by group, level, and order in the cell's list of objects.