Added "current object group" to CellDocument.
There is exactly one WorldObjectGroup active at any time for each CellDocument.
It starts off as the world's null object group.

The current group is changed when selecting a single object/group/level in
the ObjectsDock, and whenever an object becomes the only one selected (using
the object selection tool, for example).

A toolbutton in the statusbar displays the current group and allows it to be
changed via a popup menu.

When creating objects, the object's group is set to the current group.