Added object labels to CellScene.
The ObjectLabelItem class draws a background rectangle plus the name of an
object.

Each ObjectItem has a child ObjectLabelItem.

ObjectLabelItem ignores transformations (i.e., view scale) to keep itself
the same size regardless of the zoom factor.

Clicking in an object's label has the same effect as clicking the object
itself when the ObjectTool is active.  This makes it much easier to select
small objects when the view is zoomed out.

Object labels can be shown/hidden using the 'Show Object Names' action in
the View menu.