MapImages will be recreated if source maps originally had missing tilesets.
Since maps may now be successfully loaded when tileset images are missing,
map images may be created using the "missing" tileset; the binary map-image
data-file now remembers if that was the case.  They next time the data-file
is read, it will be considered invalid if the missing flag is set.  This will
cause the map image to be recreated in case the tileset issue was resolved.