MapBaseXML.txt is not a gutted .tmx anymore, it only has a list of layers.
MapBaseXML.txt used to be a .tmx file with all the desired tilesets plus
the layers (without any tile layer data).  When BMPtoTMX ran, it just filled
in the tile layer data as a comma-separated string.

One big problem with this was that the path to each tileset image was
hard-coded as ../Tiles/sometileset.png.  The user is permitted to generate
BMPtoTMX .tmx files in *any* directory, so tilesets might not be found when
opening the .tmx in TileZed.  WorldEd would find the tilesets by magic if
the Tiles directory happened to be in its list of search paths.

Another problem was how the user could add new tilesets to be included in
the generated .tmx files.

The solution is to have MapBaseXML.txt just have a list of layers that should
be in generated .tmx files.  All the tilesets are taken from Tilesets.txt.
The Tiles directory is retrieved from TileZed's settings BUT that only works
when those settings are in the Windows registry.  Gotta fix this on Mac/Unix.

A related change: the generate-lots dialog no longer requires the user to
specify the Tilesets.txt file  TileMetaInfoMgr::setTilesDirectory made
no sense, since it was specifying the directory containing Tilesets.txt and
not the Tiles directory containing the tilesets.  The location of Tilesets.txt
is assumed to be ~/.TileZed.