Added Tiles Directory to the preferences, removed search paths.
WorldEd had a list of directories to search for tilesets and maps, which is
ass-faced.  It makes to sense to "go looking" for tilesets or maps when their
locations should be exactly specified or relative to some known file or directory.
This behavior was only added to deal with the fact that TileZed did not save
the location of lot maps, it only saved the file name (without extension either).
Now that TileZed was fixed to save the location of lot maps this list of search
path is defunct.