Added QtLockedFile class to prevent simultaneous read/write of .tmx and .tbx files.
This class was taken from QtCreator.  It provides a sort of mutex lock on
file reads and writes that works across threads and processes as long as
each thread/process uses the QtLockedFile class.

Simultaneous reads are possible, but only one writer is allowed.  Any attempt
to lock a file for reading or writing while it is being written will block
the new readers/writers until writing is complete.  Simultaneous reading
and writing is not allowed.