Revised the thread-related classes.
BaseWorker now takes an InterruptibleThread argument in its constructor.

Access to mInterrupted goes through a mutex (except where MapRenderer is
accessing it).

A QWaitCondition is used to sleep the app thread while waiting for a busy
worker to finish, rather than looping and calling msleep(1).

The mWorkPending code used by various worker object classes was moved to
the BaseWorker class.