Rooms in a .lotheader have a list of rectangles making up each room.
The .lotheader file already had a list of buildings, and for each building
a list of room rectangles making up rooms in that building.  Every room
rectangle was made into a separate room, which wasn't correct for non-rectangular
rooms.

Now, each room has its own list of one or more adjacent room rectangles
that make up that room.  Adjacent rectangles are merged only when the names
of the rectangles contain a hash # character and are the same.