Added mouse-over highlighting of Lots that are part of a cell's .tmx file.
Lots added in WorldEd have a scene object to make it obvious they can be
edited.  But Lots that are part of a .tmx don't have a scene object.  Because
all Lots may be right-clicked for "Open in TileZed" action, I added a mouse-over
highlight to .tmx lots when using SubMapTool.