Really stop MapManager processing maps recursively.
When a thread finishes loading a map, it sends a signal which is delivered
via the application's event loop.  One problem with this is it can happen
any time the application processes events.  Sometimes the progress dialog
is put up while waiting for a map to load, while calling qApp->processEvents()
to keep the app responsive.