Added WorldScript.
Each world has a list of scripts applied in order. Each script has a list
of key/value parameters and 0 or more paths that the script will operate on.
The scripts are responsible for putting tiles into the world's tile layers.
All tile-drawing functions will operate in *world coordinates*. Tiles put in
by scripts are not going to be saved anywhere, the scripts should always
produce the same tiles.

Scripts have a region() function that returns the tile region *in world coordinates*
that the script effects. Signed 32-bit ints are what QRegion uses and should
be big enough for any world we'll be creating.  The regions will be used to
quickly determine which scripts must be re-run to generate tiles when other
scripts are added/removed/changed in a given area.