Added TilePainter, WorldTileLayer, WorldTileset, WorldTile.
World now has lists of tilesets and tile layers.  The WorldTileLayer class
doesn't store any data at the moment.  Instead, it uses the abstract TileSink
class to put tiles somewhere.  CellScene currently implements its own sink
that puts tiles into the map.

The TilePainter class is like QPainter, it has a bunch of methods for
filling rectangles, regions, and paths with tiles.  It can stroke a path
to a certain thickness then fill that with tiles.