Added ShadowWorld.
The path scenes display a shadow copy of the actual world. This is similar
to the way BuildingEd works.  It allows modification of the world visually
during editing without touching the original world data.  Plus the original
world can be undergoing changes (while scripts are running etc) without the
display getting into a bad state which was a big issue in TileZed.