Added SelectMovePathTool.  Re-run scripts in the area that nodes are moved.
Nodes keep a list of paths using them so that when a node moves, WorldLookup
can update the bounds of paths using those nodes.

When nodes are moved, WorldChunkMap will find any scripts in the area of the
moved nodes (original and updated positions) and run their region() function.
It takes the old + new regions and erases all the tiles in the combined region.
Then it runs the run() function of any script in the erased region to regenerate
the tiles.