Call CompositeLayerGroup::synch() in CompositeLayerGroupItem::paint().
When a map changed on disk, MapComposite would call CompositeLayerGroup::setNeedsSynch(true)
as way of preventing CompositeLayerGroupItem() from painting the layer group
while the map was being reloaded.  This was fragile, and had recently been
changed so that affected layergroups cleared the list of map layers to prevent
accessing deleted map layers.

CompositeLayerGroupItem::paint() now calls synch() when needed, which fixes
flashing (such as when undoing adding a lot).