Rewrote BMPToTMX to use BmpBlender.
BmpBlender was based on the code in BMPToTMX, but it was never a good idea
to have the same code in 2 places.  Plus BmpBlender is a lot faster due to
recent changes.

Fixed reportUknownColors() not doing anything due to mImages being cleared
right before calling it.

The unknown-colors code now correctly checks for unknown colors in each BMP
image separately.  If a color in the "veg" image corresponded to a "main" rule
but no "veg" rule with that color existed (or vice versa) it wouldn't be
reported.