Fixed (hopefully) the minimap not showing in cell views sometimes.
This bug started happening after updating to Qt 5.15.2.

There were "unable to set geometry on qwidgetwindow" errors in the console
issued in the setViewport() call in BaseGraphicsView::setUseOpenGL().

This fix seems to be adding "this" as parent to the QGLWidget constructor.

Turing "hardware accelerated rendering" on/off seems to break the minimap
sometimes.